requirements.txt
Flask==3.0.3
Flask[async]==3.0.3
flask_cors==4.0.1
python-telegram-bot==21.4

